Modified EventKey to be public
